chore: test namespace has been renamed
