--- tests/Global/App/Exceptions/AuthenticationExceptionTest.php
@@ -1,28 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests\Global\App\Exceptions;
-
-use App\Exceptions\AuthenticationException;
-use Lion\Exceptions\Exception;
-use Lion\Request\Http;
-use Lion\Request\Status;
-use Lion\Test\Test;
-
-class AuthenticationExceptionTest extends Test
-{
-    /**
-     * @throws Exception
-     * @throws AuthenticationException
-     */
-    public function testAuthenticationException(): void
-    {
-        $this
-            ->exception(AuthenticationException::class)
-            ->exceptionMessage('ERR')
-            ->exceptionStatus(Status::SESSION_ERROR)
-            ->exceptionCode(Http::HTTP_UNAUTHORIZED)
-            ->expectLionException();
-    }
-}
